refactor(api): move workspace_context to api
