+++ apps/backend/app/domains/media/api/media_router.py
@@ -6,13 +6,13 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, File, HTTPException, UploadFile
 
 from app.api.deps import get_current_user
+from app.api.workspace_context import require_workspace
 from app.core.deps import get_storage
 from app.core.log_events import (
     node_cover_upload_fail,
     node_cover_upload_start,
     node_cover_upload_success,
 )
-from app.core.workspace_context import require_workspace
 from app.domains.media.application.ports.storage_port import IStorageGateway
 from app.domains.media.application.storage_service import StorageService
