+++ tests/unit/test_node_visibility_invalidation.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import types
 import uuid
 
@@ -9,7 +11,7 @@ from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 from sqlalchemy.orm import sessionmaker
 
 from app.api import deps as api_deps
-from app.core.workspace_context import require_workspace
+from app.api.workspace_context import require_workspace
 from app.domains.nodes.api.nodes_router import router as nodes_router
 from app.domains.nodes.infrastructure.models.node import Node
 from app.domains.quests.infrastructure.models.navigation_cache_models import (
