+++ tests/unit/test_nodes_manage_router.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import importlib
 import sys
 import types
@@ -14,8 +16,8 @@ from sqlalchemy.orm import sessionmaker
 app_module = importlib.import_module("apps.backend.app")
 sys.modules.setdefault("app", app_module)
 
+import app.api.workspace_context as ws_ctx  # noqa: E402
 from app.api import deps as api_deps  # noqa: E402
-from app.core import workspace_context as ws_ctx  # noqa: E402
 from app.domains.navigation.api.nodes_manage_router import (  # noqa: E402
     router as manage_router,  # noqa: E402
 )
